finagle-http2: Server shouldn't require httpCompressor
Problem

Http2CleartextServerInitializer adds H2Filter before the httpCompressor role
in the netty pipeline, but the httpCompressor role doesn't always get added.
In particular, when the compression level is set to 0, we skip adding it
entirely.

Solution

Change from adding it before the httpCompressor role to instead adding it after
the Http2MultiplexCodec#0 role, which is the same position, but the multiplex
codec gets added unconditionally.

JIRA Issues: CSL-7376

Differential Revision: https://phabricator.twitter.biz/D251320