finagle-core: Change ServerBuilder to Act Like ClientBuilder
Problem / Solution

`ServerBuilder` does not maintain a reference to the underlying
`Stack` implementation except through means of a function which
constructs a `Server`. This means that extensions to `ServerBuilder`
are incapable of manipulating the underlying `Stack` (to add a `Stack`
module as an example). `ClientBuilder` allows this. With the change to
remove a deprecated `stack` method from `ServerBuilder`, `ServerBuilder`
can too. Change `ServerBuilder` to hold onto the underlying `Stack`
implementation directly and work more like `ClientBuilder`.

JIRA Issues: CSL-7384

Differential Revision: https://phabricator.twitter.biz/D252142