finagle-zipkin-core: use earliest span timestamp
Problem

Spans are assigned a timestamp at the time of converting to thrift,
using the minimum of the current time and the timestamps of the
constituent annotations. A span is converted to thrift once it is
likely complete, according to heuristics that rely on certain
timestamped annotations (ClientReceive/ServerSend). When a span is
missing those annotations, it will be held in the DeadlineSpanMap for
up to ttl seconds (default 120), and if there is no other timestamped
annotation to use a timestamp from, will then be assigned a time of
now. As a result, the span timestamp can be tens of seconds after the
actual span creation time.

Solution

Optionally pass creation timestamp to the Span constructor and
use if minimal.

Result

All spans are timestamped correctly.

JIRA Issues: CSL-7258

Differential Revision: https://phabricator.twitter.biz/D252767