finagle-core: Have `StackBasedServer` Extend `Stack.Transformable`
Problem

`StackBasedClient` extends `Stack.Transformable` while `StackBasedServer`
does not. With recent changes to `ServerBuilder` to maintain a hold on a
`StackBasedServer` instance, it would be beneficial to be able to manipulate
the `Stack` via that `StackBasedServer` instance.

Solution

Have `StackBasedServer` extend `Stack.Transformable`, which also brings it
into parity with `StackBasedClient` on this front.

JIRA Issues: CSL-7390

Differential Revision: https://phabricator.twitter.biz/D253542