finagle-core: Extend WireTracingFilter to the server
Problem

Finagle servers do not annotate traces with WireReceive or WireSend annotations.

Solution

Add WireReceive and WireSend annotations as close to the bottom as possible for
the StackServer. Move the ServerReceive and ServerSend annotations close to the
listener.

Result

Finagle servers will generate traces with WireReceive, ServerReceive,
ServerSend, and WireSend annotations, in that order.

JIRA Issues: CSL-7201

Differential Revision: https://phabricator.twitter.biz/D246503