finagle-core: Classify Failed Writes As ChannelClosedExceptions
Problem / Solution

Some failed writes are incorrectly classified as `UnknownChannelException`s,
when they should be classified as `ChannelClosedException`s. This is due to a
missing entry for an expected response when a peer closes a connection. Add the
proper response to 'ChannelClosedStrings' so that failed writes are seen
properly as `ChannelClosedException`s.

JIRA Issues: CSL-7399

Differential Revision: https://phabricator.twitter.biz/D256007