finagle-core: Make Transport Address Methods Final
Problem / Solution

The `Transport.localAddress` and `Transport.remoteAddress` methods
have been deprecated in favor of the respective methods on `TransportContext`
for over a year. They cannot, however, be removed at this time. This commit
provides a default implementation which points at `TransportContext` and marks
the `Transport` methods as final.

JIRA Issues: CSL-7365

Differential Revision: https://phabricator.twitter.biz/D256257