finagle-netty4: Allow Internal Extension of SharedChannelStats
Problem

In some cases, it may be necessary or desirable to track
additional client or server wide metrics which are not accounted
for in the `ChannelStatsHandler`'s `SharedChannelStats`.

Solution

Make `SharedChannelStats` extensible within Finagle. Then provide
a `Stack.Param` which is a constructor function for swapping out the
implementation of `SharedChannelStats` which is used for a particular
client or server.

JIRA Issues: CSL-6941

Differential Revision: https://phabricator.twitter.biz/D258247