finagle-core: fix concurrency bug in WatermarkPool#close
Problem

Under high concurrency, the load balancer would get into a
bad state because of a `ConcurrentModificationException` thrown
in the `close` method of `WatermarkPool`. The pool was accessing
`waiters` concurrently because `toSeq` proxies to the underlying
data structure (rather than creating a new one).

Solution

Call `toArray` on `waiters` to get a new collection. Additionally,
catch exceptions in `close` in the balancer so that we don't enter
a bad state.

JIRA Issues: CSL-7468

Differential Revision: https://phabricator.twitter.biz/D261349