finagle-thriftmux, scrooge: Clients now trace rpc method name
Problem

ThriftMux clients did not trace the method name of their RPC.
This was difficult in part because the "thrift" code happens
outside of the Finagle stack which has tracers properly
initialized.

Solution

Expand the scope of `ClientDeserializeCtx` to also include
a spot for the method name. Then read this value as part of
the ThriftMux Finagle stack.

JIRA Issues: CSL-7574

Differential Revision: https://phabricator.twitter.biz/D269383