finagle-core: return an empty balancer in partition instead of nothing
Problem

TrafficDistributor constructs balancers from endpoints. Sometimes endpoints is
empty, which we handle as a special case by returning a balancer with an empty
address set. If endpoints is empty in two consecutive updates, we end up
removing the "empty" balancer and returning nothing.

To illustrate why this happens, imagine we have a balancer set (1, 2, 3) and
receive an empty endpoint set. At the end of the scanLeft,

   balancers.keySet -- weightedGroups.keySet // Set(1, 2, 3) -- Set() == Set(1, 2, 3)

Folding over `merged` to remove (1, 2, 3) leaves us with balancer 0 for the
next iteration. Upon receiving the second empty endpoint set, we compare again
but now `balancers.keySet == Set(0)` which ends up removing the "empty"
balancer.

Solution

Restructure the code so that adding an "empty" balancer is the last thing it
does if the result would otherwise be an empty set.

JIRA Issues: CSL-7509

Differential Revision: https://phabricator.twitter.biz/D262960