finagle-thriftmux: Trace negative serialization timing values as errors
Problem

We are seeing cases in production where de-serialization times are, but
serialization times are not, being traced. This is due to a negative
serialization timing value but do not have any other information regarding what
that value is because it is discarded.

Solution

Trace negative timing values as an error annotation so we can debug this issue.
If the value is Time.Bottom it is possible that there is a race condition and
the value is not being set on the object we think we should be setting it on,
any other value would indicate that we are somehow incorrectly calculating how
we time these operations.

JIRA Issues: CSL-7584

Differential Revision: https://phabricator.twitter.biz/D274708