finagle-core: Service.pending converts a Future of Service to a Service
Problem

Finagle Stack modules compose into ServiceFactories. Some APIs (e.g. Finatra)
are Service-oriented, composed with Filters.

It's possible to convert a Service into a ServiceFactory, which then can make
use of Stack module functionality. But there is no way to convert back to a
Service (well, there is FactoryToService, but it's not what we want here).

Solution

We can get a Future[Service] from the ServiceFactory. And Future[Service] can
be a reasonable implementation of Service.

Requests wait behind resolution of the future. If resolution fails, all
requests to this service reflect the same failure. While resolution is pending,
the status of the service is Busy, otherwise it's whatever the underlying
service reports.

Closing the service is uninterruptible and instant: failing all subsequent
requests, and interrupting those in flight. Close returns Future.Done always,
independent of the underlying service.

    val svc = Service.pending(Future.never)
    val rep = svc(1)
    val closed = svc.close()

    Await.result(closed) // Future.Done
    Await.result(rep)    // Exception: Service is closed

Note: Because queued requests rely on queue order of future callbacks,
it's not guaranteed that requests to the underlying service will be
received in the same order they were sent.

JIRA Issues: CSL-7673

Differential Revision: https://phabricator.twitter.biz/D280401