finagle-http,thriftmux: Tracing annotation for backup requests
Problem

It is valuable to know in traces when a server is processing a request
that was initiated by a backup request. There is no indiciation of
this in traces today.

Solution

Add a binary annotation, "srv/backup_request_processing", to HTTP and
ThriftMux servers that indicates this.

JIRA Issues: CSL-7753

Differential Revision: https://phabricator.twitter.biz/D285486