finagle-core: Rename `SslSessionInfo` and Make It Public
Problem / Solution

`SslTransportContext` was added to provide users with more detailed
information about the SSL/TLS session being used by an existing connection.
We would like to start using it, but also don't like the name. Let's rename
it to `SslSessionInfo` and make it public, so that it can be used as part
of Finagle's API.

Differential Revision: https://phabricator.twitter.biz/D286242