finagle-core: Change Transport to Have Robust SSL/TLS Session Info
Problem

The `Transport` (via the `TransportContext`) and the `PushChannelHandle`
provides access to the peer certificate associated with the connection
if the connection is using SSL/TLS. It is desirable to have access to more
information from the SSL/TLS session when the underlying connection is using
SSL/TLS.

Solution

Change the `TransportContext` and `PushChannelHandle` to work with an
`SslSessionInfo` instead of a peer certificate directly. This provides
more direct information, and also access to the `SSLSession` being used by the
underlying connection.

Result

Eventually the `Transport.peerCertificate` Finagle context can be deprecated and
a new API can be added where the SSL/TLS session information is provided instead.
Also this information can be more actively tracked and monitored.

Differential Revision: https://phabricator.twitter.biz/D285926