finagle-core: Make maxDepth in Namer configurable
Problem

The current check for recursion level is 100 in lookup. This number is
high and could potentially cause memory and garbage collection issues.

Solution

We are making the maxDepth variable configurable instead of fixed at
100. This will allow us to change the value in the future.

Result

The recursion level check will be configurable through a global
variable which is set through environment variables. The default for
maxDepth will be 100.

JIRA Issues: FES-6046

Differential Revision: https://phabricator.twitter.biz/D286444