finagle-core: Add a way to convert an HDR histogram to BucketAndCounts
Problem

We'd like to be able to represent the WindowedPercentileHistogram in JSON. The
JSON representation will later be sent to an admin endpoint and aggregated with
other histograms.

Solution

I added a new method that converts the current histogram to a sequence of
BucketAndCount. I've also added a test to exercise this behavior.

Result

With this change, we can convert a WindowedPercentileHIstogram to a format that
can be easily serialized to JSON.

Differential Revision: https://phabricator.twitter.biz/D262166