finagle-core: Change iterator logic in WPH to allow for serialization
Problem

Currently, when we try to serialize a sequence of BucketAndCount it fails
because of the iterator logic (the HDR Histogram's custom iterator doesn't play
well with the jackson object mapper).

Solution

The solution is to return a HDR histogram specific iterator.

Result

Serialization to json works as a result.

Differential Revision: https://phabricator.twitter.biz/D290721