finagle: Tags stack param
Problem / Solution

Tags are a more powerful Label.

Tags associate Finagle clients and servers with a set of keywords. Labels
are simply Tags with a single keyword.

Tags provide a general purpose configuration mechanism for functionality
that is not yet known. This is powerful, but also easily misused. As such,
be conservative in using them.

Frameworks that create services for each endpoint should tag them with
endpoint metadata, e.g.,

    val showPost = Http.server.withLabels("GET", "/posts/show")
    val createPost = Http.server.withLabels("POST", "PUT", "/posts/create")

Differential Revision: https://phabricator.twitter.biz/D289540