finagle-http: Strip overhead from trailers when they are empty
Problem

Additional `HeaderMap` is allocated for each HTTP message to accommodate trailers, even
if they are empty (the majority of the cases).

Solution

We can strip these allocations and reduce message's footprint for certain types of messages:
inbound chunked messages, inbound full messages that don't have trailers, and all outbound
messages.

Result

We only pay (at the cost of allocations) for trailers when they are present. If there are no trailers
on the wire, they are free.

JIRA Issues: CSL-7827

Differential Revision: https://phabricator.twitter.biz/D292530