finagle-netty4: Add pending_io_events gauge
Problem

There is no visibility into the workload that's queued up in the Netty's pipeline (in the IO
thread pool) before it enters the Finagle space. This kind of overload scenario typically
arises when the thread (event-loop) servicing a multiplexed connection is blocked so it
can no longer process new IO events.

Solution

Add a new metric, `pending_io_events` that exports the number of pending tasks in the
Netty's event-loops servicing the given client or server. The new metric (a gauge) allows
to quickly identify a source of a slowdown in Finagle applications.

Additionally, the new metric can point to a "bad neighbor" when a client/server slowdown
is caused by a misbehaving sibling client/server (running in the same IO thread pool)
overloading an IO thread.

JIRA Issues: CSL-3586

Differential Revision: https://phabricator.twitter.biz/D293376