finagle-mysql: Introduce MysqlTransporter
Problem

MySQL session establishment currently occurs in the `connPhase` of
the `ClientDispatcher`. However, it should occur instead earlier as
part of the session acquisition phase within a Transporter. This will
also be the proper place to implement SSL encrypted session establishment
also.

Solution

Encapsulate the use of the `Netty4Transporter` within a new `MysqlTransporter`,
which will also act as an extension point for switching where finagle-mysql does
session establishment for the MySQL protocol.

JIRA Issues: CSL-7176

Differential Revision: https://phabricator.twitter.biz/D294790