finagle-mysql: Break Out Settings from Handshake
Problem

The `Handshake` class contains both settings for a MySQL
session as well as behavior for how to create one. This is
undesirable, as we want to change how and where the session
is created.

Solution

Separate out the settings from the behavior. Place the settings
in a new `HandshakeSettings` class.

JIRA Issues: CSL-7874

Differential Revision: https://phabricator.twitter.biz/D294836