finagle-mysql: Move LostSyncException Out of ClientDispatcher
Problem / Solution

LostSyncException is a generic exception within finagle-mysql which
indicates that communication with the server has been corrupted and
both sides are no longer in sync. With the effort to move where
session establishment takes place for finagle-mysql, it will be
possible for this to happen outside of the `ClientDispatcher`, and
so this functionality should be usable outside of the `ClientDispatcher`.
Let's move it.

JIRA Issues: CSL-7878

Differential Revision: https://phabricator.twitter.biz/D295425