finagle-mysql: MysqlTransporter Should Work with Packets
Problem / Solution

The `MysqlTransporter` follows the encapsulated `Netty4Transporter` in providing
a `Transport` which is capable of reading and writing `Buf`s. However, for
MySQL, we immediately modify the created `Transport` to instead work with
`Packet`s. Let's change the `MysqlTransporter` to instead return a `Transport`
directly which works with `Packet`s.

JIRA Issues: CSL-7900

Differential Revision: https://phabricator.twitter.biz/D295978