finagle-mysql: Introduce MysqlTransport
Problem / Solution

We are moving MySQL session establishment from the domain of the
`ClientDispatcher` to the domain of the `MysqlTransporter` and the
created `Transport`. Having a specific `MysqlTransport` which knows
about `Packet`s and how to properly `close` a MySQL session allows
us greater flexibility in placing functionality in the appropriate
locations. It will also allow for future manipulation of the underlying
Netty channel during SSL/TLS handshaking. So, let's add it.

JIRA Issues: CSL-7902

Differential Revision: https://phabricator.twitter.biz/D296902