finagle-core: Optimize map lookups in BindingFactory
Problem

Profiling shows an unexpected amount of time in the 3
`ServiceFactoryCaches` in `BindingFactory`.

Solution

Introduce a mixin `CacheHashCode` that lazily computes the `hashCode`
and returns that calculation on subsequent calls.

Result

Less time spent on `hashCode` in `BindingFactory` at the cost of an
extra member on a few classes.

JIRA Issues: CSL-7791

Differential Revision: https://phabricator.twitter.biz/D296662