finagle-mysql: Move ConnectionPhase to Handshake
Problem

Right now, handshaking for finagle-mysql only supports a plain handshake, and
only within the `ClientDispatcher`. We would like to move the handshaking to
be done as part of service acqusition, and also be able to add support for
SSL/TLS.

Solution

Remove the particulars of the handshaking from the `ClientDispatcher`, and move
them into a reworked `Handshake` class.

JIRA Issues: CSL-7917

Differential Revision: https://phabricator.twitter.biz/D298400