finagle-core, twitter-server: Add `c.t.finagle.DtabFlags`
Problem

There is currently no defined way for users of the `c.t.finagle.Dtab` API to
append to the `Dtab.base` delegation table with a supplemental value that is
passed at runtime.

Solution

Add the `c.t.finagle.DtabFlags` trait which defines a `c.t.app.Flag`, "dtab.add"
that can provide a supplemental delegation table (`Dtab`) to append to the
current server `Dtab.base`.

The trait exposes a `addDtabs()` function which applies the parsed Flag value
and appends it to the current `Dtab.base`.

Result

Users are able to augment the base delegation table (`Dtab.base`) of the server
with a value passed at runtime via the command line.

JIRA Issues: CSL-7766

Differential Revision: https://phabricator.twitter.biz/D297596