finagle-core: Reordered field so it can't be referenced before it's initialized
Problem

We pass a call-by-name parameter when we initialize a field that references
another field that hasn't been initialized yet.  This is legal because it's a
call-by-name parameter, so it could in theory used safely, and it typically is,
but we've received reports that there are occasionally NullPointerExceptions.

Solution

Initialize the field before we construct the call-by-name parameter that refers
to it.

JIRA Issues: CSL-7849

Differential Revision: https://phabricator.twitter.biz/D301661