finagle-mysql: Add Toggle for Moving Handshaking
Problem

MySQL session establishment (via the handshaking in the connection phase)
should occur as part of service acquisition within the transporter. Instead
today it occurs as part of the client dispatcher. We would prefer session
establishment to happen when creating the transport, and feel that it is
necessary to do so when adding SSL/TLS, as SSL/TLS requires manipulation of
the underlying Netty pipeline, and it is easier to do that via the
transporter and transport.

Solution

Add a toggle for moving handshaking out of the client dispatcher and into
the transporter.

JIRA Issues: CSL-7938

Differential Revision: https://phabricator.twitter.biz/D301456