finatra-http: Get rid of MediaType
Problem

MediaType is one of our last guava dependencies, so we'll be able to start
removing guava once we can get rid of it.

Solution

Replace MediaType with a String everywhere the finatra API requires a MediaType,
and rely on method in finagle's MediaType to check equality for media types and
handle utf8 charsets.

JIRA Issues: CSL-5672

Differential Revision: https://phabricator.twitter.biz/D308761