scrooge-generator: Provide Implementation for ServicePerEndpoint.filtered
Problem / Solution

Classes which extend `ServicePerEndpoint` are not provided with a working
implementation of its `filtered` method. This can be confusing for users
who attempt to apply a TypeAgnostic filter and find that it does't work
without overriding the `filtered` method to provide a concrete implementation.
Compounding the problem is that using `ServicePerEndpoint.apply` does return
an instance with a working `filtered` method. So, let's delegate the
implementation of the extension to the companion object.

JIRA Issues: CSL-8042

Differential Revision: https://phabricator.twitter.biz/D309920