finagle-core: Expose the TraceId context Key
Problem

We use pub/sub libraries for many purposes here. Unfortunately, when a
traced request ends at a publisher, there is no way to relate that trace
to the subscriber that picks up the event. In order to accomplish that,
we'd need to be able to serialize the current trace context along with
the event payload, then deserialize it at the subscriber and set the
appropriate context for the request. To do that, we need access to the
`Context.broadcast.Key[T]` instance that finagle uses to serialize the
context.

Solution

This change exposes that Key by renaming `idCtx` to `TraceIdContext` on the
`c.t.f.tracing.Trace` object and making it `private[twitter]`.

Result

The `Context.broadcast.Key[TraceId]` is now accessible to things in the
`com.twitter` package.

JIRA Issues: IOP-386

Differential Revision: https://phabricator.twitter.biz/D313547