finagle-http: Remove Misleading MaxRequest/MaxResponse Size Methods
Problem

The `MaxRequestSize` param within Finagle HTTP is applicable to Finagle
HTTP servers only. The `MaxResponseSize` param is applicable to Finagle
HTTP clients only. It is incorrect and misleading to have a
`client.withMaxRequestSize(size)` method and a
`server.withMaxResponseSize(size)` method, since while they change the
underlying `Stack` params for that client or server, changing that `Stack`
param for that particular side has no effect.

Solution

Remove the `client.withMaxRequestSize(size)` and
`server.withMaxResponseSize(size)` methods, and more clearly document the
underlying params to indicate that `MaxRequestSize` is for HTTP servers only,
and `MaxResponseSize` is for HTTP clients only.

JIRA Issues: CSL-8121

Differential Revision: https://phabricator.twitter.biz/D314019