finagle: Add ryano to OWNERS
Problem / Solution

I have made numerous changes to the core components and protocols of Finagle,
and rewritten entire sections. I thoroughly understand how all the pieces fit
together at this point, and already am an owner of most of them. It's time to
add myself as a top level owner.

Differential Revision: https://phabricator.twitter.biz/D315494