finagle-netty4: Make Client and Server Components Work with LocalAddress
Problem / Solution

Netty's `LocalChannel` is a great avenue for being able to test `Channel` and
`ChannelPipeline` related functionality without requiring full on sockets. It
sits in the spot between an `EmbeddedChannel` and a full on `Channel`. Let's
extend Finagle's `ConnectionBuilder` and `ListeningServerBuilder` to support it
via a Netty `LocalAddress`.

JIRA Issues: CSL-8038

Differential Revision: https://phabricator.twitter.biz/D315320