finagle-core: Treat failures consistently when resolving & binding union paths
Problem:
The name resolver ignores failures when resolving unions whereas the
name binder treats a single failure as terminal. This inconsistency is
dangerous - if a single resolution in a union flaps between successful and
 failed, the entire serverset will be emptied out leading to 100% failure.

For example, with the path (/a & /b), if /b flaps into a failed binding
state, the overall resolution will flap into failed and the serverset
will empty out.

Solution:
Both name resolution and binding should filter out failures when building a
union. In the example above, if /b flaps to a failed binding,
the union will resolve to just /a.

Differential Revision: https://phabricator.twitter.biz/D315282