finagle-core: Fix logic for Aperture.dapertureActive
Problem

`Aperture.dapertureActive` does not include a check for whether or not
a `ProcessCoordinate` is defined. As such, the result of
`dapertureActive` is not accurate.

`D303833` caused a regression as a result of this for services without
a coordinate.  The load band would no longer change offered size and
thus be fixed.

Solution

Include the `ProcessCoordinate` check in `dapertureActive`.

Result

This allows `LoadBand` to have the correct check, allowing the load
band to expand properly when there isn't a `ProcessCoordinate`.

JIRA Issues: CSL-8078

Differential Revision: https://phabricator.twitter.biz/D315028