[finagle-thrift] Add the trace_id_high field to Span.
Problem:

Zipkin can use a 128b ID. It does this by adding a second 64b id to each span.
This was included in the `RequestHeader` struct, but not in the `Span` struct.

Solution:

Add the missing field to the `Span` struct.

Result:

The `Span` struct now has an optional `trace_id_high` field.

JIRA Issues: CSL-8204

Differential Revision: https://phabricator.twitter.biz/D317847