finagle-mysql: Close prepared statements when removed from cache
Problem

In a system that uses a high number of prepared statements, but has low
gc,  finagle-mysql would not eagerly close PreparedStatements on the server.
This  could result in the database server running out of PreparedStatements.
finagle-mysql caches PreparedStatements and enforces a limit on the number
of  prepared statements that can be cached at any one time. Releasing
the  PreparedStatement on the database server was only done after
the  PreparedStatement is no longer referenced, and gc occurs.

The need to release PreparedStatements after all references were released was
because `PreparedStatement.apply` within a transaction loses the checkout
semantics guaranteed by the DefaultPool since we replace the factory for a
client with a singleton factory. This is problematic since we can now interleave
`PrepareRequest`s which can subsequently evict `PrepareOk` results we've
returned to the application.

Solution

finagle-mysql now releases PreparedStatements immediately, when ejected from
the  cache.  Introduce an async mutex within `session` (the building block for
`transaction`) which allows the underlying service to be used exclusively until
a call to `close`, matching the semantics in `DefaultPool`.

Result

The maximum number of PreparedStatements open on the database server
is  guaranteed to be # of connections * PreparedStatement cache size.  Only a
single level of nested transactions are supported. Attempting to nest  more than
that will result in an IllegalStateException.  Any support for nested
transactions may be removed at a later time. Use of  nested transactions with
finagle-mysql is discouraged as it is not supported by  MySql and is unlike to
work as the developer expects.

JIRA Issues: CSL-4741

Differential Revision: https://phabricator.twitter.biz/D318568