finagle: Add statsFilter for streaming
Problem

The Finagle Http StatsFilter is only enabled with `withHttpStats`, however, we
need all stream related stats to be populated as long as `isChunked` is set to
true in request or response.

Solution

Have a StatsFilter dedicated for stream metrics.

Result

Created a StreamingStatsFilter, moved stream duration metrics from
`http/StatsFilter` to it, added counters for closed streams, total streams, and
stream failures.

Note: We are collecting stream failures metrics in `HttpClientDispatcher` and
`HttpServerDispatcher` as `stream/failures/<exception_name>` and
`stream/failures`. However, we don't distinguish a request stream failure from a
response stream failure. Also, maintaining stats in a dispatcher is expensive as
the stats are initialized for every connection. The plan is to deprecate these
counters after this change is approved.

JIRA Issues: CSL-8041

Differential Revision: https://phabricator.twitter.biz/D315041