Workaround for Rsc mixed compile public/private scoping
Problem

In preparation of supporting compilation of finagle with Rsc,
a manual workaround needs to be merged regarding
a public/private scoping inconsistency.

Solution

Remove the "private" modifier of PostRequestEvidence.

Result

We will be able to apply an automatic rewrite to the rest of
finatra/http/src/test/scala for Rsc compatiblility.

Differential Revision: https://phabricator.twitter.biz/D322570