finagle-http2: Disable ping-based failure detector
Problem

There is something funky about how ping-based failure detector is wired in H2 that
it does more harm than good: when/if it kicks in, the client is having a really hard time
to recover. Disabling it and relying on other circuit breakers in the stack instead, makes
client way more stable under a high load.

Solution

Disable ping-based failure detector in H2 until we better understand what's
wrong with it.

JIRA Issues: CSL-8039

Differential Revision: https://phabricator.twitter.biz/D322357