finagle-mysql: Don't try and close a connection that is already closed
Problem:
If a connection is already closed, the RollbackFactory will also attempt
to poison the connection and close it.  This causes alarming spam in the logs.

Solution:
If the rollback fails due to a ChannelClosedException, don't do anything else.

Result:
An already closed connection, that has an attempted rollback, doesn't result
in a log message that says the close() call failed.

JIRA Issues: DINF-1081

Differential Revision: https://phabricator.twitter.biz/D322588