finagle-netty4: Add BufCodecPipeline
Problem / Solution

Creating a Finagle protocol based on `Buf` is the most straightforward
way for creating and testing a Finagle protocol. To do so, this generally
requires creating a function capable of manipulating a Netty pipeline to add
`BufCodec` so that `ByteBuf`s can be converted to and from Twitter `Buf`s.
Let's add something which does this generically within Finagle.

Differential Revision: https://phabricator.twitter.biz/D322731