finagle-zipkin-core: Address race condition in DeadlineSpanMap
Problem

DeadlineSpanMap uses a mutable structure, `MutableSpan`, to buffer up
changes for a given TraceId. It has an async task, `flush`, that runs
periodically and can cause calls to `update` to be dropped.

Solution

Use locking to check when this occurs and if it does, run the update
operation again and immediately run `logSpans` on that.

Result

No more lost updates.

JIRA Issues: CSL-8210

Differential Revision: https://phabricator.twitter.biz/D319367