finagle-core: Add SSL/TLS Session Information to ClientConnection
Problem

We would like to be able to export and monitor more information about
whether particular connections are using SSL/TLS, and if so, the
specifics about that connection. For servers, we track connections
based on information contained within the `ClientConnection` class.

Solution

Add `SslSessionInfo` to `ClientConnection` so that server connections
can track whether connections are using SSL/TLS and the specifics
related to a secure connection.

Differential Revision: https://phabricator.twitter.biz/D323305