finagle-http: Properly set Transport.peerCertificate for HTTP/2
Problem

The `ChannelTransportContext` is responsible for retrieving
`SslSessionInfo` from a Netty4 `Channel`. It does so by extracting
the `SslHandler` out of the `ChannelPipeline`, and if it exists, uses
the `SSLEngine` contained within to construct an `SslSessionInfo`
appropriately.

However, when using a Netty multiplexed protocol, such as HTTP/2, the
`Channel` that is processing the request is not the one where we want
to be looking for the `SslHandler`. Instead, that should be in the
parent `Channel`s `ChannelPipeline`.

Solution

Modify the `ChannelTransportContext` to check the current
`Channel` for an `SslHandler`. If found, use it, otherwise check
the `Channel`s parent until one is found, or the chain ends.

JIRA Issues: CSL-8249

Differential Revision: https://phabricator.twitter.biz/D324392