finagle-core: UsingSslSessionInfo Fails When Local Certificates are Null
Problem / Solution

`UsingSslSessionInfo` takes in an `SSLSession` and calls the session's
`getLocalCertificates` method to retrieve the collection of local
certificates sent as part of this `SSLSession`. However, that method
can return 'null', and when it does, construction of `UsingSslSessionInfo`
fails. Add a guard which checks for 'null'.

JIRA Issues: CSL-8262

Differential Revision: https://phabricator.twitter.biz/D324499